+++ test/cube_index_tests.c
@@ -6,7 +6,7 @@
 Test(cube_index, c_index_is_section) {
   for (uint32_t i = 0; i < 10000; i++) {
     cube_t c = c_index_to_cube(i);
-    uint32_t j = cube_to_c_index(&c, UD);
+    uint64_t j = cube_to_c_index(&c, UD);
     cr_assert_eq(i, j);
   }
 }
